# Contracts/openapi/development-sync-v1.yaml
openapi: 3.1.0
info:
  title: Virtual Indoor Rowing development sync API
  version: 1.0.0
  description: Loopback-only development infrastructure; uploaded sessions are never ranked or authoritative race results.
servers:
  - url: http://127.0.0.1:8080
paths:
  /v1/development/bootstrap:
    post:
      summary: Exchange a Compose-generated bootstrap secret for a 24-hour development bearer token.
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/BootstrapRequest'}}}}
      responses: {'200': {description: Token issued, content: {application/json: {schema: {$ref: '#/components/schemas/BootstrapResponse'}}}}, default: {$ref: '#/components/responses/Problem'}}
...
components:
  securitySchemes:
    developmentBearer: {type: http, scheme: bearer, bearerFormat: opaque}
...
  responses:
    Problem: {description: RFC 9457 problem, content: {application/problem+json: {schema: {$ref: '#/components/schemas/Problem'}}}}
  schemas:
    BootstrapRequest: {type: object, required: [request_version, bootstrap_secret], properties: {request_version: {type: integer, const: 1}, bootstrap_secret: {type: string, minLength: 32}}}
    BootstrapResponse: {type: object, required: [response_version, development_identity, bearer_token, expires_at], properties: {response_version: {type: integer, const: 1}, development_identity: {type: string}, bearer_token: {type: string}, expires_at: {type: string, format: date-time}}}
...
    Problem: {type: object, required: [type, title, status], properties: {type: {type: string, format: uri-reference}, title: {type: string}, status: {type: integer}, detail: {type: string}, instance: {type: string}}}